requirements.txt
python-telegram-bot
python-dotenv
ccxt
pandas
mplfinance
requests
httpx